Create Fall 2022 COP 1334 A2: Video Game Info
The purpose of this assignment was to create a system for tracking the games and the scores in terms of graphics and replay value, and then compute the highest score in each category and the highest overall score. The main point of this assignment was to implement what we learned about arrays.