Create F22 COP 1047C A1: Basic input output
The purpose of this assignment was to create basic code that could receive input, calculate with said input, and output the results. The main point of this assignment was to implement what we learned about input/output and working with variables.

NOTE: This was my first assignment for this course.